LayoutTests/http/tests/security/local-CSS-from-remote-expected.txt

CONSOLE MESSAGE: line 0: Not allowed to load local resource: file:///tmp/LayoutTests/http/tests/security/resources/cssStyle.css
This test is to see if a remote file can include a local CSS style. 
If the background is yellow then the CSS was loaded.

Test Passed.
